fix(harness): harden hook installation
