+++ tooling/arnes/src/measure/install/validate/cursor.rs
@@ -18,6 +18,7 @@ const EVENTS: &[&str] = &[
     "afterFileEdit",
     "beforeSubmitPrompt",
     "preCompact",
+    "postCompact",
     "stop",
     "afterAgentResponse",
     "afterAgentThought",
